+++ str/ft_strlcat.c
@@ -6,7 +6,7 @@
 /*   By: gpoblon <gpoblon@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/09/23 17:35:41 by gpoblon           #+#    #+#             */
-/*   Updated: 2016/09/23 17:44:04 by gpoblon          ###   ########.fr       */
+/*   Updated: 2016/09/27 12:14:40 by gpoblon          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_strlcat(char *dest, const char *src, size_t size)
 {
-	int i;
+	size_t i;
 
 	i = 0;
 	while (src[i++]);
